Rock Class and Delegate
I'm attempting to move all the logic and properties into a Rock class.

I'm having trouble getting the delegation to work. It is hitting the method where the rock image is touched the method isn't being sent to the viewController through delegation.

Also, should we have the Rock class be a model and just store the data of each property while the View Controller handles the logic of if the Rock is old enough for stuff? Or should all of that logic be in the Rock class?